Create erc20
Solidity version: The first line specifies the Solidity compiler version (^0.8.0) this code is compatible with.
SPDX license identifier: This line indicates the license under which the code is distributed (MIT License in this case).
Import statement: We import the ERC20 standard functionalities from the OpenZeppelin library (@openzeppelin/contracts/token/ERC20/ERC20.sol). This saves us from writing repetitive code for token transfers, balances, etc.
Contract definition: The GoogolBase contract inherits from the ERC20 contract provided by OpenZeppelin.
Token details:
name: This variable stores the token's name ("GoogolBase Token").
symbol: This variable stores the token's symbol ("GOGL").
totalSupply: This variable stores the total number of tokens minted.
Constructor: The constructor function is called when the contract is deployed. It takes the initial supply of tokens (initialSupply) as an argument:
Calls the inherited ERC20 constructor to set the token name and symbol.
Assigns the initial supply to the totalSupply variable.
Mints the initial supply of tokens to the address that deployed the contract (msg.sender).